Exception registration moved
The function flask_sandman.app._register_error_handlers was moved to flask_sandman.exception.register
